requirements.txt
allure-pytest
pytest
python-dotenv
requests
selene==2.0.0rc4
